Move vtkCutter logic from SolidActor to AnalysisActor
